+++ app/src/main/java/labs/lucka/mlp/MainActivity.kt
@@ -1,28 +1,25 @@
 package labs.lucka.mlp
 
-import android.app.ActivityManager
-import android.content.Context
 import android.content.Intent
-import android.location.Criteria
-import android.location.LocationManager
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import android.support.design.widget.Snackbar
 import android.support.v4.widget.NestedScrollView
-import android.support.v7.app.AlertDialog
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
 import android.view.Menu
 import android.view.MenuItem
-import android.view.View
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.content_main.*
-import kotlinx.android.synthetic.main.dialog_add_mock_target.view.*
-import java.io.*
 
 /**
  * MainActivity for MLP
+ *
+ * ## Changelog
+ * ### 0.2
+ * - Replace saveData and loadData with [DataKit]
+ * - Migrate isServiceOnline] and isMockLocationEnabled to [MockLocationProviderService]
+ * - Migrate showDeveloperOptionsDialog and showAddMockTargetDialog to [DialogKit]
  *
  * ## Private Attributes
  * - [mockTargetList]
@@ -33,13 +30,7 @@ import java.io.*
  * ### Overridden
  * - [onCreate]
  * ### Private
- * - [saveData]
- * - [loadData]
  * - [updateFabService]
- * - [isMLPServiceOnline]
- * - [isMockLocationEnabled]
- * - [showDeveloperOptionsDialog]
- * - [showAddMockDialog]
  *
  * @author lucka-me
  * @since 0.1
@@ -56,7 +47,11 @@ class MainActivity : AppCompatActivity() {
         object : MainRecyclerViewAdapter.MainRecyclerViewListener {
 
             override fun onRemovedAt(position: Int) {
-                saveData()
+                try {
+                    DataKit.saveData(this@MainActivity, mockTargetList)
+                } catch (error: Exception) {
+                    DialogKit.showSimpleAlert(this@MainActivity, error.message)
+                }
             }
 
         }
@@ -66,7 +61,12 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         setSupportActionBar(toolbar)
 
-        loadData()
+        //loadData()
+        try {
+            mockTargetList = DataKit.loadData(this)
+        } catch (error: Exception) {
+            DialogKit.showSimpleAlert(this@MainActivity, error.message)
+        }
 
         mainRecyclerViewAdapter =
             MainRecyclerViewAdapter(this, mockTargetList, mainRecyclerViewListener)
@@ -74,7 +74,22 @@ class MainActivity : AppCompatActivity() {
         mainRecyclerView.adapter = mainRecyclerViewAdapter
 
         fabAddMockTarget.setOnClickListener { _ ->
-            showAddMockDialog()
+            DialogKit.showAddMockTargetDialog(
+                this,
+                mockTargetList,
+                {
+                    Snackbar
+                        .make(mainRecyclerView, R.string.err_coordinate_wrong, Snackbar.LENGTH_LONG)
+                        .show()
+                },
+                {
+                    mainRecyclerViewAdapter.notifyAddMockTarget()
+                    Snackbar.make(
+                        nestedScrollView,
+                        R.string.add_mock_target_success,
+                        Snackbar.LENGTH_LONG
+                    ).show()
+                })
         }
 
         nestedScrollView.setOnScrollChangeListener(
@@ -90,13 +105,17 @@ class MainActivity : AppCompatActivity() {
         updateFabService()
 
         fabService.setOnClickListener { _ ->
-            if (isMLPServiceOnline()) {
+            if (MockLocationProviderService.isServiceOnline(this)) {
                 val mlpService = Intent(this, MockLocationProviderService::class.java)
                 stopService(mlpService)
                 updateFabService()
             } else {
-                saveData()
-                if (isMockLocationEnabled()) {
+                try {
+                    DataKit.saveData(this@MainActivity, mockTargetList)
+                } catch (error: Exception) {
+                    DialogKit.showSimpleAlert(this@MainActivity, error.message)
+                }
+                if (MockLocationProviderService.isMockLocationEnabled(this)) {
                     val mlpService =
                         Intent(this, MockLocationProviderService::class.java)
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
@@ -106,7 +125,7 @@ class MainActivity : AppCompatActivity() {
                     }
                     updateFabService()
                 } else {
-                    showDeveloperOptionsDialog()
+                    DialogKit.showDeveloperOptionsDialog(this)
                 }
             }
         }
@@ -138,175 +157,17 @@ class MainActivity : AppCompatActivity() {
     }
 
     /**
-     * Save the [mockTargetList] to file.
-     *
-     * @author lucka-me
-     * @since 0.1
-     */
-    private fun saveData() {
-        val filename = getString(R.string.data_filename)
-        val file = File(filesDir, filename)
-        val fileOutputStream: FileOutputStream
-        val objectOutputStream: ObjectOutputStream
-
-        try {
-            fileOutputStream = FileOutputStream(file)
-            objectOutputStream = ObjectOutputStream(fileOutputStream)
-            objectOutputStream.writeObject(mockTargetList)
-            objectOutputStream.close()
-            fileOutputStream.close()
-        } catch (error: Exception) {
-            DialogKit.showSimpleAlert(this, error.message)
-        }
-    }
-
-    /**
-     * Load the [mockTargetList] from file.
-     *
-     * @author lucka-me
-     * @since 0.1
-     */
-    private fun loadData() {
-        val filename = getString(R.string.data_filename)
-        val file = File(filesDir, filename)
-        val fileInputStream: FileInputStream
-        val objectInputStream: ObjectInputStream
-
-        if (!file.exists()) return
-
-        try {
-            fileInputStream = FileInputStream(file)
-            objectInputStream = ObjectInputStream(fileInputStream)
-            @Suppress("UNCHECKED_CAST")
-            mockTargetList = objectInputStream.readObject() as ArrayList<MockTarget>
-            objectInputStream.close()
-            fileInputStream.close()
-        } catch (error: Exception) {
-            DialogKit.showSimpleAlert(this, error.message)
-        }
-    }
-
-    /**
      * Update the icon of [fabService]
      *
      * @author lucka-me
      * @since 0.1.1
      */
     private fun updateFabService() {
-        fabService.setImageDrawable(
-            if (isMLPServiceOnline()) getDrawable(R.drawable.ic_cancel)
-            else getDrawable(R.drawable.ic_start)
-        )
+        fabService.setImageDrawable(getDrawable(
+            if (MockLocationProviderService.isServiceOnline(this)) R.drawable.ic_cancel
+            else R.drawable.ic_start
+        ))
     }
 
-    /**
-     * Used to detect if the [MockLocationProviderService] is running.
-     *
-     * ## Changelog
-     * ### 0.1.1
-     * - Use [ActivityManager.getRunningServices] instead of PreferenceManager
-     *
-     * @return True if is running, false if not.
-     * @author lucka-me
-     * @since 0.1
-     */
-    private fun isMLPServiceOnline(): Boolean {
-        val activityManager = getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
-        @Suppress("DEPRECATION")
-        for (serviceInfo in activityManager.getRunningServices(Int.MAX_VALUE)) {
-            if (serviceInfo.service.className == MockLocationProviderService::class.java.name)
-                return true
-        }
-        return false
-    }
 
-    /**
-     * Used to detect if the Enabled mock location is on and available for MLP.
-     *
-     * @return True if the option is on and available for MLP, false if off or unavailable.
-     *
-     * @author lucka-me
-     * @since 0.1.2
-     */
-    private fun isMockLocationEnabled(): Boolean {
-        val testLocationManager = getSystemService(Context.LOCATION_SERVICE) as LocationManager
-        try {
-            testLocationManager.addTestProvider(
-                getString(R.string.test_location_provider),
-                false, false,false, false,
-                true,true, true,
-                Criteria.POWER_LOW, Criteria.ACCURACY_FINE
-            )
-        } catch (error: SecurityException) {
-            showDeveloperOptionsDialog()
-            return false
-        }
-        testLocationManager.removeTestProvider(getString(R.string.test_location_provider))
-        return true
-    }
-
-    /**
-     * Show a dialog to explain the Enable mock location option
-     * and provide a button to open the Developer Options.
-     *
-     * @author lucka-me
-     * @since 0.1.2
-     */
-    private fun showDeveloperOptionsDialog() {
-        DialogKit.showDialog(
-            this,
-            R.string.mock_location_option_title,
-            R.string.mock_location_option_text,
-            positiveButtonTextId = R.string.settings,
-            positiveButtonListener = { _, _ ->
-                try {
-                    startActivity(Intent(Settings.ACTION_APPLICATION_DEVELOPMENT_SETTINGS))
-                } catch (error: Exception) {
-                    DialogKit.showSimpleAlert(
-                        this,
-                        R.string.err_developer_options_failed
-                    )
-                }
-            },
-            negativeButtonTextId = R.string.cancel,
-            cancelable = false
-        )
-    }
-
-    /**
-     * Show a dialog to add a new mock target.
-     *
-     * @author lucka-me
-     * @since 0.1
-     */
-    private fun showAddMockDialog() {
-        val dialogLayout = View.inflate(this, R.layout.dialog_add_mock_target, null)
-        AlertDialog.Builder(this)
-            .setTitle(R.string.add_mock_target_title)
-            .setView(dialogLayout)
-            .setPositiveButton(R.string.confirm) { _, _ ->
-                val longitude = dialogLayout.longitudeEdit.text.toString().toDoubleOrNull()
-                val latitude = dialogLayout.latitudeEdit.text.toString().toDoubleOrNull()
-                if (longitude == null || latitude == null ||
-                    longitude < -180 || longitude > 180||
-                    latitude < -90 || latitude > 90
-                ) {
-                    Snackbar
-                        .make(mainRecyclerView, R.string.err_coordinate_wrong, Snackbar.LENGTH_LONG)
-                        .show()
-                } else {
-                    mockTargetList.add(MockTarget(longitude, latitude))
-                    saveData()
-                    mainRecyclerViewAdapter.notifyAddMockTarget()
-                    Snackbar.make(
-                        nestedScrollView,
-                        R.string.add_mock_target_success,
-                        Snackbar.LENGTH_LONG
-                    ).show()
-                }
-            }
-            .setNegativeButton(R.string.cancel, null)
-            .setCancelable(false)
-            .show()
-    }
 }