+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 <h1 align=center>Changelog</h1>
 
 ```markdown
+## [0.2] - 2018-09-02
+- 0.1.4(115) -> 0.2(144)
+- New data structure
+
+### Added
+- New attributes for MockTarget:
+  - Title (UI and methods updated)
+  - Inteval, accuracy and altitude (UI and methods not updated yet)
+- Save / load data with the new DataKit and Gson
+
+### Changed
+- MockTarget is a data class now, the old data could NOT be migrated
+```
+
+```markdown
 ## [0.1.4] - 2018-09-02
 - 0.1.2(62) -> 0.1.4(115)
 - New feature: Preference, enable/disable providers